chore: Streamline rule table for embedded builds
Refactor rule table and dependencies to reduce binary size and
complexity for embedded environments by creating lighter-weight/stub
implementations and isolating them via build tags.
