+++ internal/engine/tracer/sink_common.go
@@ -1,0 +1,40 @@
+// Copyright 2021-2025 Zenauth Ltd.
+// SPDX-License-Identifier: Apache-2.0
+
+package tracer
+
+import (
+	"sync"
+
+	enginev1 "github.com/cerbos/cerbos/api/genpb/cerbos/engine/v1"
+)
+
+type Sink interface {
+	Enabled() bool
+	AddTrace(trace *enginev1.Trace)
+}
+
+type Collector struct {
+	traces []*enginev1.Trace
+	mutex  sync.RWMutex
+}
+
+func NewCollector() *Collector {
+	return &Collector{}
+}
+
+func (c *Collector) Enabled() bool {
+	return true
+}
+
+func (c *Collector) AddTrace(trace *enginev1.Trace) {
+	c.mutex.Lock()
+	defer c.mutex.Unlock()
+	c.traces = append(c.traces, trace)
+}
+
+func (c *Collector) Traces() []*enginev1.Trace {
+	c.mutex.RLock()
+	defer c.mutex.RUnlock()
+	return c.traces
+}
