+++ internal/observability/logging/signal.go
@@ -1,7 +1,7 @@
 // Copyright 2021-2025 Zenauth Ltd.
 // SPDX-License-Identifier: Apache-2.0
 
-//go:build !windows
+//go:build !windows && !js && !wasm
 
 package logging
