+++ internal/util/globs_embedded.go
@@ -1,0 +1,8 @@
+// Copyright 2021-2025 Zenauth Ltd.
+// SPDX-License-Identifier: Apache-2.0
+
+//go:build js && wasm
+
+package util
+
+func logError(string, error) {}
